refactor: move addBaseUrl to utils-common
