+++ packages/docusaurus-utils-common/src/__tests__/isInternalUrl.test.ts
