+++ packages/docusaurus-utils-common/src/isInternalUrl.ts
